backend/requirements.txt
# This file is auto-generated from requirements.in
# To update, run: pip-compile requirements.in

# Django Framework
Django==5.2

# API Framework - Django Shinobi (fork of Django Ninja)
django-shinobi

# Database Driver
psycopg2-binary==2.9.9

# Development & Debugging Tools
django-extensions==3.2.3

# Additional useful packages
python-decouple==3.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3

# JWT Authentication
PyJWT==2.9.0
djangorestframework-simplejwt==5.3.1

# Celery for background tasks
celery==5.4.0
redis==5.1.1
django-redis==5.4.0
django-celery-beat==2.9.0

# Production server
gunicorn==23.0.0

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0

# Code quality
black==24.8.0
flake8==7.1.1

# Additional dependencies that might be required
pillow==10.4.0
python-dateutil==2.9.0
pytz==2024.1
asgiref==3.8.1
sqlparse==0.5.1

